DRY up some stuff
